Refactor: 피드 form 텍스트 입력 렌더링 최적화 (#17)
